Added controller file, updated route file
