+++ package.json
@@ -1,0 +1,28 @@
+{
+  "name": "Node_Authentication",
+  "version": "1.0.0",
+  "description": "Guide for making authentication API utilizing json web tokens",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/gregdbanks/Node_Authentication.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/gregdbanks/Node_Authentication/issues"
+  },
+  "homepage": "https://github.com/gregdbanks/Node_Authentication#readme",
+  "dependencies": {
+    "bcryptjs": "^2.4.3",
+    "dotenv": "^14.3.2",
+    "express": "^4.17.2",
+    "express-validator": "^6.14.0",
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^6.1.8"
+  }
+}
