+++ routes/user.js
@@ -1,17 +1,10 @@
 const express = require("express");
-const { check, validationResult } = require("express-validator");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const router = express.Router();
+const { check } = require("express-validator");
+
 const auth = require("../middleware/auth");
+const { signUp, loginUser, getMe } = require("../controllers/user");
 
-const User = require("../model/User");
-
-/**
- * @method - POST
- * @param - /signup
- * @description - User SignUp
- */
+const router = express.Router();
 
 router.post(
   "/signup",
@@ -22,60 +15,7 @@ router.post(
       min: 6,
     }),
   ],
-  async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        errors: errors.array(),
-      });
-    }
-
-    const { username, email, password } = req.body;
-    try {
-      let user = await User.findOne({
-        email,
-      });
-      if (user) {
-        return res.status(400).json({
-          msg: "User Already Exists",
-        });
-      }
-
-      user = new User({
-        username,
-        email,
-        password,
-      });
-
-      const salt = await bcrypt.genSalt(10);
-      user.password = await bcrypt.hash(password, salt);
-
-      await user.save();
-
-      const payload = {
-        user: {
-          id: user.id,
-        },
-      };
-
-      jwt.sign(
-        payload,
-        "randomString",
-        {
-          expiresIn: 10000,
-        },
-        (err, token) => {
-          if (err) throw err;
-          res.status(200).json({
-            token,
-          });
-        }
-      );
-    } catch (err) {
-      console.log(err.message);
-      res.status(500).send("Error in Saving");
-    }
-  }
+  signUp
 );
 
 router.post(
@@ -86,67 +26,9 @@ router.post(
       min: 6,
     }),
   ],
-  async (req, res) => {
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        errors: errors.array(),
-      });
-    }
-
-    const { email, password } = req.body;
-    try {
-      let user = await User.findOne({
-        email,
-      });
-      if (!user)
-        return res.status(400).json({
-          message: "User Not Exist",
-        });
-
-      const isMatch = await bcrypt.compare(password, user.password);
-      if (!isMatch)
-        return res.status(400).json({
-          message: "Incorrect Password !",
-        });
-
-      const payload = {
-        user: {
-          id: user.id,
-        },
-      };
-
-      jwt.sign(
-        payload,
-        "randomString",
-        {
-          expiresIn: 3600,
-        },
-        (err, token) => {
-          if (err) throw err;
-          res.status(200).json({
-            token,
-          });
-        }
-      );
-    } catch (e) {
-      console.error(e);
-      res.status(500).json({
-        message: "Server Error",
-      });
-    }
-  }
+  loginUser
 );
 
-router.get("/me", auth, async (req, res) => {
-  try {
-    // request.user is getting fetched from Middleware after token authentication
-    const user = await User.findById(req.user.id);
-    res.json(user);
-  } catch (e) {
-    res.send({ message: "Error in Fetching user" });
-  }
-});
+router.get("/me", auth, getMe);
 
 module.exports = router;
